requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pygame>=2.0.0
pymunk>=5.7.0
pillow>=8.0.0
pydantic>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
opencv-python>=4.5.0
PyExecJS>=1.5.0
cffi>=1.15.0
